fake up waitpid() for windows
